Feat: adicionando os para nao recriar csv e keyboardinterrupt para nao dar erro

+++ Python/data-collector.py
@@ -7,6 +7,7 @@ try:
 except ImportError:
     pynvml = None
 import csv
+import os
 
 def capture(components):
     user = get_mac_address()
@@ -146,9 +147,12 @@ def store(data):
     with open('data.csv', 'a', newline='') as csvfile:
         writer = csv.writer(csvfile, delimiter=';')
         writer.writerow([data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7], data[8], data[9], data[10], data[11], data[12], data[13], data[14]])
-
-with open('data.csv', 'w', newline='') as csvfile:
-        writer = csv.writer(csvfile, delimiter=';')
-        writer.writerow(["user", "cpu_percent", "cpu_frequency", "ram_percent", "swap_memory_total", "swap_memory_used", "swap_memory_percent", "upload_speed", "download_speed", "temperature", "fans_speed", "disk", "gpu_usage", "gpu_energy", "timestamp"])
-
-capture([1,1,1,1,1,1,1,1,1,1,1,1,1])
+if(not os.path.exists('data.csv')):
+    with open('data.csv', 'w', newline='') as csvfile:
+            writer = csv.writer(csvfile, delimiter=';')
+            writer.writerow(["user", "cpu_percent", "cpu_frequency", "ram_percent", "swap_memory_total", "swap_memory_used", "swap_memory_percent", "upload_speed", "download_speed", "temperature", "fans_speed", "disk", "gpu_usage", "gpu_energy", "timestamp"])
+try:
+    capture([1,1,1,1,1,1,1,1,1,1,1,1,1])
+except KeyboardInterrupt:
+    print("Encerrado")
+
